+++ attendancemanagementsystem/src/main/java/com/example/attendancemanagementsystem/classroom/classplace/service/MdClassroomService.java
@@ -1,4 +1,4 @@
-package com.example.attendancemanagementsystem.user.admin.service;
+package com.example.attendancemanagementsystem.classroom.classplace.service;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -11,18 +11,17 @@ import com.example.attendancemanagementsystem.common.entity.ClassroomEntity;
 import com.example.attendancemanagementsystem.common.repository.ClassroomRepository;
 
 @Service
-public class AdminClassroomService {
+public class MdClassroomService {
 
     @Autowired
     private ClassroomRepository classroomRepository;
 
-    // --- 全件取得 ---
+    // 全件取得 
     public List<ClassroomEntity> getAllClassrooms() {
         return classroomRepository.findAll();
     }
 
-    // --- 一括保存・削除処理 ---
-    // ★引数に macAddresses を追加
+    // 一括保存・削除処理 
     @Transactional
     public void saveClassroomList(
             List<Integer> classroomIds, 
@@ -31,7 +30,7 @@ public class AdminClassroomService {
         
         if (classroomNames == null) return;
 
-        // 1. 削除処理（画面から消されたIDを特定して削除）
+        // 削除処理
         List<Integer> keptIds = new ArrayList<>();
         if (classroomIds != null) {
             for (Integer id : classroomIds) {
@@ -46,11 +45,10 @@ public class AdminClassroomService {
             }
         }
 
-        // 2. 更新・新規登録処理
+        // 更新・新規登録処理
         for (int i = 0; i < classroomNames.size(); i++) {
             String name = classroomNames.get(i);
             Integer currentId = (classroomIds != null && classroomIds.size() > i) ? classroomIds.get(i) : null;
-            // ★MACアドレスを取得
             String mac = (macAddresses != null && macAddresses.size() > i) ? macAddresses.get(i) : null;
 
             ClassroomEntity entity;
@@ -61,7 +59,7 @@ public class AdminClassroomService {
             }
 
             entity.setClassroomName(name);
-            entity.setMacAddress(mac); // ★セット
+            entity.setMacAddress(mac); 
             
             classroomRepository.save(entity);
         }
