+++ attendancemanagementsystem/src/main/java/com/example/attendancemanagementsystem/classroom/subject/dto/MasterDataResponse.java
@@ -1,0 +1,9 @@
+package com.example.attendancemanagementsystem.classroom.subject.dto;
+
+import java.util.List;
+
+public record MasterDataResponse(
+    List<String> subjects,
+    List<String> courses,
+    List<String> classrooms
+) {}
