+++ attendancemanagementsystem/src/main/java/com/example/attendancemanagementsystem/user/admin/dto/MasterDataResponse.java
@@ -1,9 +1,9 @@
-package com.example.attendancemanagementsystem.user.admin.dto;
+// package com.example.attendancemanagementsystem.user.admin.dto;
 
-import java.util.List;
+// import java.util.List;
 
-public record MasterDataResponse(
-    List<String> subjects,
-    List<String> courses,
-    List<String> classrooms
-) {}
+// public record MasterDataResponse(
+//     List<String> subjects,
+//     List<String> courses,
+//     List<String> classrooms
+// ) {}
